+++ lib/whois/record/parser/whois.rotld.ro.rb
@@ -14,23 +14,21 @@ module Whois
   class Record
     class Parser
 
-      #
-      # = whois.rotld.ro parser
-      #
       # Parser for the whois.rotld.ro server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
+      # @note This parser is just a stub and provides only a few basic methods
+      #   to check for domain availability and get domain status.
+      #   Please consider to contribute implementing missing methods.
+      #
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
       class WhoisRotldRo < Base
 
         property_supported :status do
           if content_for_scanner =~ /Domain Status:\s(.+?)\n/
             case $1.downcase
-            when "ok"
+            when "ok", "updateprohibited"
               :registered
             else
               Whois.bug!(ParserError, "Unknown status `#{$1}'.")