+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - NEW: Added full whois.comlaude.com parser (GH-222). [Thanks @delwyn] 
 
+- NEW: Added #domain, #domain_id to whois.tcinet.ru.
+
 - FIXED: ARIN queries required additional params (GH-220, GH-10). [Thanks @linrock]
 
 - FIXED: Fixed ARPA Reverse DNS lookup.